+++ src/npc/aggression.rs
@@ -11,7 +11,10 @@ use crate::{
     tablet::mind_control::MindControled,
 };
 
-use super::NPC;
+use super::{
+    movement::{NewDirectionEvent, ResetAggroEvent},
+    NPC,
+};
 
 // :0
 
@@ -123,5 +126,23 @@ pub fn add_pursuit_urge(
                 target.0 = Some(ev.target_entity);
             }
         }
+    }
+}
+
+/// The npc returns to walk peacefully
+///
+/// - Remove ChaseBehavior
+/// - Insert WalkBehavior
+/// - Ask for a new destination
+pub fn reset_aggro(
+    mut commands: Commands,
+    mut reset_aggro_event: EventReader<ResetAggroEvent>,
+
+    mut new_direction_event: EventWriter<NewDirectionEvent>,
+) {
+    for ev in reset_aggro_event.iter() {
+        commands.entity(ev.npc).remove::<ChaseBehavior>();
+        commands.entity(ev.npc).insert(WalkBehavior);
+        new_direction_event.send(NewDirectionEvent(ev.npc));
     }
 }
